refactor: optiomize RMQ declarer
